ingestion-service/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
kafka-python==2.0.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9